:recycle: Move JavaScript parser into its own file.
